refactor(RedisInstance): move aws client to common package
