+++ cmd/main.go
@@ -39,10 +39,10 @@ import (
 	"github.com/kyma-project/cloud-manager/pkg/quota"
 
 	"github.com/kyma-project/cloud-manager/pkg/common/abstractions"
+	awsclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/client"
 	awsiprangeclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/iprange/client"
 	awsnfsinstanceclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/nfsinstance/client"
 	awsnukenfsclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/nuke/client"
-	awsredisinstanceclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/redisinstance/client"
 	awsvpcpeeringclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/vpcpeering/client"
 	azureiprangeclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/azure/iprange/client"
 	azurenetworkclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/azure/network/client"
@@ -303,7 +303,7 @@ func main() {
 		mgr,
 		gcpmemorystoreclient.NewMemorystoreClientProvider(),
 		azureredisclient.NewClientProvider(),
-		awsredisinstanceclient.NewClientProvider(),
+		awsclient.NewElastiCacheClientProvider(),
 		env,
 	); err != nil {
 		setupLog.Error(err, "unable to create controller", "controller", "RedisInstance")
