+++ pkg/kcp/provider/aws/mock/elastiCacheClientFake.go
@@ -13,8 +13,8 @@ import (
 	secretsmanagerTypes "github.com/aws/aws-sdk-go-v2/service/secretsmanager/types"
 	"github.com/elliotchance/pie/v2"
 	"github.com/google/uuid"
+	awsclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/client"
 	awsmeta "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/meta"
-	awsclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/redisinstance/client"
 	"k8s.io/utils/ptr"
 )
