+++ pkg/kcp/provider/aws/mock/type.go
@@ -4,7 +4,6 @@ import (
 	awsclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/client"
 	iprangeclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/iprange/client"
 	nfsinstanceclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/nfsinstance/client"
-	redisinstanceclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/redisinstance/client"
 	vpcpeeringclient "github.com/kyma-project/cloud-manager/pkg/kcp/provider/aws/vpcpeering/client"
 	scope "github.com/kyma-project/cloud-manager/pkg/kcp/scope/client"
 )
@@ -29,7 +28,7 @@ type Clients interface {
 	IpRangeClient
 	NfsClient
 	VpcPeeringClient
-	redisinstanceclient.ElastiCacheClient
+	awsclient.ElastiCacheClient
 }
 
 type Providers interface {
@@ -37,7 +36,7 @@ type Providers interface {
 	IpRangeSkrProvider() awsclient.SkrClientProvider[iprangeclient.Client]
 	NfsInstanceSkrProvider() awsclient.SkrClientProvider[nfsinstanceclient.Client]
 	VpcPeeringSkrProvider() awsclient.SkrClientProvider[vpcpeeringclient.Client]
-	ElastiCacheProviderFake() awsclient.SkrClientProvider[redisinstanceclient.ElastiCacheClient]
+	ElastiCacheProviderFake() awsclient.SkrClientProvider[awsclient.ElastiCacheClient]
 }
 
 type Configs interface {
